snapshot

+++ bind9/setup.sh
@@ -3,9 +3,9 @@
 source $D/../bin/base.sh
 source $D/.env
 
-if istrue BIND9_INSTALL_SRC; then
-
+if istrue BIND9_COMPILE_SRC; then
     preinstall_apt pkg-config python3-pip libuv1 libuv1-dev libssl-dev libcap-dev libxml2-dev
+    
     cd `check git bind9 https://github.com/isc-projects/bind9.git  -b v9_16_9 --depth=1` || exit $?
     ./configure --prefix=/usr/local/bind9 \
         --sysconfdir=/etc/bind9 \
@@ -13,9 +13,12 @@ if istrue BIND9_INSTALL_SRC; then
         --with-openssl \
         --with-libxml2
     make -j2 
+    cd -
+fi
 
-    isactive bind9 ||  $S systemctl stop bind9
-
+if istrue BIND9_INSTALL_SRC; then
+    cd `check bind9`
+    isactive bind9 &&  $S systemctl stop bind9
     $S rm -rf /usr/local/bind9
     $S make install
     cd -
@@ -33,12 +36,11 @@ if istrue BIND9_UPDATE_CONFIG; then
 
     $S mkdir -p /var/named/data
     $S mkdir -p /var/log/bind9 
-    $S mkdir -p /var/run/bind9
     # rndc.key
     $S /usr/local/bind9/sbin/rndc-confgen -a
 
     # 默认用named用户启动，配置为640，named:named 权限
-    $S chown -R named.named /usr/local/bind9/ /etc/bind9/ /var/named/ /var/log/bind9/ /var/run/bind9/  
+    $S chown -R named.named /usr/local/bind9/ /etc/bind9/ /var/named/ /var/log/bind9/
 
     # 检查配置
     /usr/local/bind9/sbin/named-checkzone  sexhansc.com  /var/named/private/sexhansc.com.zone
